+++ src/components/combo_box/combo_box.test.js
@@ -1,6 +1,7 @@
 import React from 'react';
-import { shallow } from 'enzyme';
-import { requiredProps } from '../../test';
+import { shallow, render, mount } from 'enzyme';
+import sinon from 'sinon';
+import { requiredProps, findTestSubject } from '../../test';
 
 import { EuiComboBox } from './combo_box';
 
@@ -29,7 +30,7 @@ const options = [{
 
 describe('EuiComboBox', () => {
   test('is rendered', () => {
-    const component = shallow(
+    const component = render(
       <EuiComboBox {...requiredProps} />
     );
 
@@ -38,49 +39,45 @@ describe('EuiComboBox', () => {
 });
 
 describe('props', () => {
-  test('options', () => {
+  test('options are rendered', () => {
+    // NOTE: It's tough to test this because the dropdown containing the options opens up in
+    // a portal.
     const component = shallow(
-      <EuiComboBox
-        options={options}
-        {...requiredProps}
-      />
+      <EuiComboBox options={options} />
     );
 
     expect(component).toMatchSnapshot();
   });
 
-  test('selectedOptions', () => {
+  test('selectedOptions are rendered', () => {
     const component = shallow(
       <EuiComboBox
         options={options}
         selectedOptions={[options[2], options[4]]}
-        {...requiredProps}
       />
     );
 
     expect(component).toMatchSnapshot();
   });
 
-  describe('isClearable is false', () => {
-    test('without selectedOptions', () => {
+  describe('isClearable=false disallows user from clearing input', () => {
+    test('when no options are selected', () => {
       const component = shallow(
         <EuiComboBox
           options={options}
           isClearable={false}
-          {...requiredProps}
         />
       );
 
       expect(component).toMatchSnapshot();
     });
 
-    test('with selectedOptions', () => {
+    test('when options are selected', () => {
       const component = shallow(
         <EuiComboBox
           options={options}
           selectedOptions={[options[2], options[4]]}
           isClearable={false}
-          {...requiredProps}
         />
       );
 
@@ -88,29 +85,59 @@ describe('props', () => {
     });
   });
 
-  test('singleSelection', () => {
+  test('singleSelection is rendered', () => {
     const component = shallow(
       <EuiComboBox
         options={options}
         selectedOptions={[options[2]]}
         singleSelection={true}
-        {...requiredProps}
       />
     );
 
     expect(component).toMatchSnapshot();
   });
 
-  test('isDisabled', () => {
+  test('isDisabled is rendered', () => {
     const component = shallow(
       <EuiComboBox
         options={options}
         selectedOptions={[options[2]]}
         isDisabled={true}
-        {...requiredProps}
       />
     );
 
     expect(component).toMatchSnapshot();
+  });
+});
+
+describe('behavior', () => {
+  describe('clear button', () => {
+    test('calls onChange callback with empty array', () => {
+      const onChangeHandler = sinon.spy();
+      const component = mount(
+        <EuiComboBox
+          options={options}
+          selectedOptions={[options[2]]}
+          onChange={onChangeHandler}
+        />
+      );
+
+      findTestSubject(component, 'comboBoxClearButton').simulate('click');
+      sinon.assert.calledOnce(onChangeHandler);
+      sinon.assert.calledWith(onChangeHandler, []);
+    });
+
+    test('focuses the input', () => {
+      const component = mount(
+        <EuiComboBox
+          options={options}
+          selectedOptions={[options[2]]}
+          onChange={() => {}}
+        />
+      );
+
+      findTestSubject(component, 'comboBoxClearButton').simulate('click');
+      expect(findTestSubject(component, 'comboBoxSearchInput').getDOMNode()).toBe(document.activeElement);
+    });
   });
 });